[MOB-7613] Update usages of cache
