+++ src/inapp/request.ts
@@ -1,6 +1,7 @@
 import { delMany, entries } from 'idb-keyval';
 import { GETMESSAGES_PATH, SDK_VERSION, WEB_PLATFORM } from 'src/constants';
 import { baseIterableRequest } from 'src/request';
+import { addNewMessagesToCache, getCachedMessagesToDelete } from './cache';
 import schema from './inapp.schema';
 import {
   CachedMessage,
@@ -8,7 +9,6 @@ import {
   InAppMessageResponse,
   InAppMessagesRequestParams
 } from './types';
-import { addNewMessagesToCache, getCachedMessagesToDelete } from './utils';
 
 type RequestInAppMessagesProps = {
   latestCachedMessageId?: string;
