+++ src/inapp/tests/utils.test.ts
@@ -5,14 +5,13 @@ import MockAdapter from 'axios-mock-adapter';
 import { messages } from '../../__data__/inAppMessages';
 import { baseAxiosRequest } from '../../request';
 import { srSpeak } from '../../utils/srSpeak';
-import { CLOSE_BUTTON_POSITION, CachedMessage, InAppMessage } from '../types';
+import { CLOSE_BUTTON_POSITION } from '../types';
 import {
   addButtonAttrsToAnchorTag,
   filterHiddenInAppMessages,
   filterOnlyReadAndNeverTriggerMessages,
   generateCloseButton,
   generateWidth,
-  getCachedMessagesToDelete,
   getHostnameFromUrl,
   paintIFrame,
   paintOverlay,
@@ -1034,52 +1033,6 @@ describe('Utils', () => {
       expect(rightButton.style.right).toBe('20px');
       expect(rightButton.style.left).toBe('');
       expect(rightButton.innerHTML).toBe('');
-    });
-  });
-
-  describe('Caching', () => {
-    const now = Date.now();
-    const allMessages = [...messages];
-
-    const cachedMessages: CachedMessage[] = allMessages.flatMap((msg) => [
-      [msg.messageId, msg]
-    ]);
-
-    it('should delete cached messages that are expired', () => {
-      const unexpiredMessages = allMessages.filter(
-        (msg) => msg.expiresAt > now
-      );
-      const expiredMessageIds = allMessages.reduce(
-        (allFetchedIds: string[], message) => {
-          if (message.expiresAt < now) allFetchedIds.push(message.messageId);
-          return allFetchedIds;
-        },
-        []
-      );
-
-      const messagesForDeletion = getCachedMessagesToDelete(
-        cachedMessages,
-        unexpiredMessages
-      );
-      expect(messagesForDeletion).toEqual(expiredMessageIds);
-    });
-
-    it('should delete any cached messages not included in the fetch', () => {
-      const validMessages: InAppMessage[] = [];
-      const invalidMessages: InAppMessage[] = [];
-      allMessages.forEach((msg) =>
-        msg.messageId === 'normalMessage!'
-          ? validMessages.push(msg)
-          : invalidMessages.push(msg)
-      );
-
-      const messagesForDeletion = getCachedMessagesToDelete(
-        cachedMessages,
-        validMessages
-      );
-      expect(messagesForDeletion).toEqual(
-        invalidMessages.map((msg) => msg.messageId)
-      );
     });
   });
 });
